+++ packages/starter/src/lib/package-json-preparing.spec.ts
@@ -105,10 +105,10 @@ describe('packageJsonPreparing', () => {
       const result = await prepare({ appType: 'csr', nogit: true, tester: false, testMode: true });
 
       expect(getScripts(result)).toMatchObject({
-        analyzer: 'tsx scripts.build.ts --analyzer',
-        build: 'tsx scripts.build.ts --mode=production',
+        analyzer: 'node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON scripts.build.ts --analyzer',
+        build: 'node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON scripts.build.ts --mode=production',
         lint: 'npm run lint:ts && npm run lint:code && npm run lint:styles',
-        start: 'tsx scripts.build.ts',
+        start: 'node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON scripts.build.ts',
       });
       expect(addDependenciesMock.mock.calls[0]).toEqual([
         expect.anything(),
@@ -156,7 +156,8 @@ describe('packageJsonPreparing', () => {
         types: './dist/types/index.d.ts',
       });
       expect(getScripts(result)).toMatchObject({
-        'build:example': 'node example/scripts.build.ts --mode=production',
+        'build:example':
+          'node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON example/scripts.build.ts --mode=production',
         format: 'npm run format:prettier && npm run format:code',
         lint: 'npm run lint:ts && npm run lint:code',
         production: 'npm run lint && npm run build && npm publish',
@@ -185,8 +186,8 @@ describe('packageJsonPreparing', () => {
         const result = await prepare({ appType, nogit: true, tester: true });
 
         expect(getScripts(result)).toMatchObject({
-          test: 'tsx scripts.tests.ts',
-          'test:watch': 'tsx scripts.tests.ts --watch',
+          test: 'node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON scripts.tests.ts',
+          'test:watch': 'node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON scripts.tests.ts --watch',
         });
         expect(addedGroups().slice(-2)).toEqual([
           {